Feat: add navigation from login to speaker view
