--- prototype/Utter/Utter/Views/TandemListView.swift
@@ -1,20 +1,0 @@
-//
-//  TandemListView.swift
-//  Utter
-//
-//  Created by Victor Magnus Oldensand on 2023-10-10.
-//
-
-import SwiftUI
-
-struct TandemListView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct TandemListView_Previews: PreviewProvider {
-    static var previews: some View {
-        TandemListView()
-    }
-}
