+++ prototype/Utter/Utter/resources/Assets.xcassets/AccentColor.colorset/Contents.json
