+++ prototype/Utter/Utter/resources/Assets.xcassets/AppIcon.appiconset/Contents.json
