+++ prototype/Utter/Utter/resources/Assets.xcassets/Contents.json
